+++ packages/khao-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.5.3
+
+- use nav-tag as wrapper for `khao-pagination`
+
+## 3.5.2
+
+- added default and debugging messeages to `khao-pagintaion`
+
+## 3.5.1
+
+- Use type string for `currentPage` and `totalPages` in `khao-pagination`
+
 ## 3.5.0
 
 - added `khao-pagination` and `khao-pagintaion-item` components